Move the eval user fixture out of the ux eval
eval:durability builds its fixture with createEvalUser/deleteEvalUser,
which live inside the ux eval's own directory. Removing that eval would
take the durability one down with it.

The fixture is not the ux eval's: it belongs to whoever needs a
disposable account. A rename, nothing else.
